Move timestamp test to statement.rs and add missing cases
