+++ scylla/tests/integration/main.rs
@@ -19,6 +19,7 @@ mod shards;
 mod silent_prepare_batch;
 mod silent_prepare_query;
 mod skip_metadata_optimization;
+mod statement;
 mod tablets;
 #[path = "../common/utils.rs"]
 mod utils;
